% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{compute_clustering}
\alias{compute_clustering}
\title{Model-Based Clustering}
\usage{
compute_clustering(data)
}
\description{
The optimal model according to BIC for EM initialized by hierarchical
clustering for parameterized Gaussian mixture models.
Use the \code{\link[mclust]{Mclust}} function.
}
