% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{compute_cor}
\alias{compute_cor}
\title{Correlation and Cramér's V}
\usage{
compute_cor(x, y = NULL, method = c("pearson", "kendall", "spearman",
  "cramer"), n_jobs = 1, ...)
}
\arguments{
\item{x}{A character vector, matrix or data frame.}

\item{y}{NULL (default) or a vector if x is a vector.}

\item{method}{a character string indicating which method to use. One of
"pearson" (default), "kendall", "spearman" or "cramer".}

\item{n_jobs}{numeric, the number of thread(s) to use (1 by default). NB:
n_jobs is not yet implemented.}

\item{...}{parameters to pass to \code{\link[stats]{cor}}}
}
\value{
A numeric value if \code{x} and \code{y} are vectors or a matrix if
  \code{x} is a matrix or a data frame.
}
\description{
Compute the correlation or the Cramér's V measure of association.
}
\examples{
compute_cor(mtcars)
compute_cor(mtcars, method = "spearman")
compute_cor(tea$price, tea$home, method = "cramer")
}
\seealso{
\code{\link{compute_cramer_v}}, \code{\link{plot_cor}}
}
