% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{compute_cramer_v}
\alias{compute_cramer_v}
\title{Cramér's V}
\usage{
compute_cramer_v(x, y = NULL)
}
\arguments{
\item{x}{A character vector, matrix or data frame.}

\item{y}{NULL (default) or a vector if x is a vector.}
}
\value{
A numeric value if \code{x} and \code{y} are vectors or a matrix if
  \code{x} is a matrix or a data frame.
}
\description{
Compute the Cramér's V measure of association.
}
\examples{
compute_cramer_v(tea)
compute_cramer_v(tea$price, tea$home)
}
\seealso{
\code{\link{compute_cor}}, \code{\link{plot_cor}}
}
