% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{corr2}
\alias{corr2}
\title{Correlation}
\usage{
corr2(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
a number
}
\description{
`corr2` compute the correlation of x and y if these are vectors.
}
\details{
The inputs must be numeric, and the vectors x and y are equal in
length.
}
\examples{
x <- 1:5
y <- c(3, 5, 8, 9, 13)
corr2(x, y)
}
