% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tea}
\alias{tea}
\title{Survey on tea consumption}
\format{A data frame with 300 rows and 36 categorical variables.}
\source{
\url{https://cran.r-project.org/web/packages/FactoMineR/index.html}
}
\usage{
tea
}
\description{
A survey on tea comsumption. This dataset comes from the FactoMineR package.
}
\keyword{datasets}
