% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_impute.R
\name{var_impute}
\alias{var_impute}
\title{Impute numeric variables for analysis}
\usage{
var_impute(dta)
}
\arguments{
\item{dta}{A data frame that contains numeric variables for imputation.}
}
\value{
The same data frame with numeric variables numeric
    variables imputed.
}
\description{
Impute numeric variables for analysis
}
