% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_transform.R
\name{var_transform}
\alias{var_transform}
\title{Common transformations of numeric variables for analysis.}
\usage{
var_transform(dta)
}
\arguments{
\item{dta}{A data frame that contains numeric variables for transformation.}
}
\value{
The same data frame with additional transformations for each
          numeric variables.
}
\description{
Common transformations of numeric variables for analysis.
}
\examples{
iris_new = var_transform(iris)
}
